# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Tourtect API
  version: "1.0.0"
  description: |
    Tourtect backend API — community-powered travel safety and price transparency platform.
    All responses use snake_case JSON. Times are ISO 8601 UTC.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/price-checks:
    post:
      operationId: createPriceCheck
      tags: [pricing]
      summary: Submit a price check
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceCheckRequest"
      responses:
        "200":
          description: Price insight result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceInsight"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/safety/assessments:
    post:
      operationId: createSafetyAssessment
      tags: [safety]
      summary: Submit a safety assessment request
      parameters:
        - $ref: "#/components/parameters/RequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SafetyAssessmentRequest"
      responses:
        "200":
          description: Safety assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyAssessment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
...
    RequestIdParam:
      name: X-Request-ID
      in: header
      description: Client-provided request identifier for tracing
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    UnprocessableEntity:
      description: Request body validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
...
    AlertLevel:
      type: string
      enum: [typical, elevated, high_risk, insufficient_data]
    Vertical:
      type: string
      enum: [taxi, exchange, food, tour, street_retail]
    ServiceSegment:
      type: string
      enum: [budget, standard, premium, luxury, regulated]
    VenueType:
      type: string
      enum:
        - fixed_shop
        - casual_eatery
        - street_stall
        - mobile_vendor
        - market_stall
        - attraction_concession
        - transport_vendor
        - peer_to_peer
    TransactionContext:
      type: string
      enum:
        - posted_price
        - verbal_quote
        - negotiated
        - metered
        - platform_booked
        - unsolicited_goods_service
        - donation_solicitation
    SafetyUrgency:
      type: string
      enum: [critical, urgent, non_emergency, information]
    GeoFallbackLevel:
      type: string
      enum: [exact_zone, commune, province, province_cluster, national_vertical]
    ProblemDetail:
      type: object
      required: [type, status, title]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
        status:
          type: integer
          description: HTTP status code
        title:
          type: string
          description: Short human-readable summary
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: URI reference for this specific occurrence
        trace_id:
          type: string
          description: Request trace identifier
...
    Money:
      type: object
      required: [amount_minor, currency, exponent]
      properties:
        amount_minor:
          type: string
          description: Amount in minor units as string to avoid floating-point issues
        currency:
          type: string
          minLength: 3
          maxLength: 3
        exponent:
          type: integer
    PriceCheckRequest:
      type: object
      required: [vertical, raw_item, money, unit, region_id, service_segment, venue_type, transaction_context, observed_at]
      properties:
        vertical:
          $ref: "#/components/schemas/Vertical"
        canonical_item_id:
          type: string
        raw_item:
          type: string
        money:
          $ref: "#/components/schemas/Money"
        unit:
          type: string
        region_id:
          type: string
        pricing_zone_id:
          type: string
        service_segment:
          $ref: "#/components/schemas/ServiceSegment"
        venue_type:
          $ref: "#/components/schemas/VenueType"
        transaction_context:
          $ref: "#/components/schemas/TransactionContext"
        observed_at:
          type: string
          format: date-time
        extraction_confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        user_confirmed:
          type: boolean
          default: false
    PriceInsight:
      type: object
      required: [alert_level, observed, confidence, freshness, dataset_version, trace_id, reasons]
      properties:
        alert_level:
          $ref: "#/components/schemas/AlertLevel"
        observed:
          $ref: "#/components/schemas/Money"
        reference:
          $ref: "#/components/schemas/PriceReference"
        deviation_ratio:
          type: number
          format: double
        confidence:
          type: number
          format: double
        comparison_scope:
          type: string
        freshness:
          type: string
          format: date-time
        sample_size:
          type: integer
        snapshot_version:
          type: string
        reasons:
          type: array
          items:
            type: string
        possible_benign_explanations:
          type: array
          items:
            type: string
        dataset_version:
          type: string
        trace_id:
          type: string
    PriceReference:
      type: object
      properties:
        p10_minor:
          type: string
        p50_minor:
          type: string
        p90_minor:
          type: string
        currency:
          type: string
        exponent:
          type: integer
        unit:
          type: string
        region_id:
          type: string
        pricing_zone_id:
          type: string
        service_segment:
          $ref: "#/components/schemas/ServiceSegment"
        venue_type:
          $ref: "#/components/schemas/VenueType"
        geo_fallback_level:
          $ref: "#/components/schemas/GeoFallbackLevel"
        effective_sample_size:
          type: integer
        independent_source_count:
          type: integer
    SafetyAssessmentRequest:
      type: object
      required: [observed_facts]
      properties:
        observed_facts:
          type: array
          items:
            type: string
          minItems: 1
        user_reported_state:
          type: string
        threat_indicators:
          type: array
          items:
            type: string
        injury_indicators:
          type: array
          items:
            type: string
        confinement_indicators:
          type: array
          items:
            type: string
        coercion_indicators:
          type: array
          items:
            type: string
        ability_to_leave:
          type: boolean
        user_confirmed_facts:
          type: array
          items:
            type: string
        region_id:
          type: string
          description: Explicit region for versioned safety-directory lookup; omission disables regional hotline lookup.
    SafetyAssessment:
      type: object
      required: [urgency, safe_actions, safety_directory_version, confidence, trace_id]
      properties:
        urgency:
          $ref: "#/components/schemas/SafetyUrgency"
        safe_actions:
          type: array
          items:
            type: string
        approved_action_codes:
          type: array
          items:
            type: string
        explanation_codes:
          type: array
          items:
            type: string
        silent_mode_recommended:
          type: boolean
        surface_emergency_options:
          type: boolean
        emergency_service_ids:
          type: array
          items:
            type: string
        safety_directory_version:
          type: string
        confidence:
          type: number
          format: double
        trace_id:
          type: string
...